Ejercicio11.java

+++ Ejercicio11.java
@@ -3,14 +3,14 @@ import java.util.Scanner;
 public class Ejercicio11 {
     public static void main(String[] args) {
         Scanner inputScanner = new Scanner(System.in);
-        int num;
-        int suma = 0;
-        int cantidadNum = 0;
+        double num;
+        double suma = 0;
+        double cantidadNum = 0;
 
         System.out.println("Introduce números, introduce uno negativo para calcular la media:");
 
         do {
-            num = inputScanner.nextInt();
+            num = inputScanner.nextDouble();
 
             if (num < 0) {
                 break;
